refactor (#183) : 댓글과 답글 컴포넌트 분리
