+++ src/pages/Board/FreeBoard/FreeBoard.jsx
@@ -14,6 +14,7 @@ import { fetchPostItems } from "../../../services/api/postApi";
 import { useDispatch } from "react-redux";
 import { setAllPostItems } from "../../../store/post/postSlice";
 import { BarLoading } from "../../../components/Common/LodingSpinner";
+import Comment from "../../../components/Common/Comment/Comment";
 
 function FreeBoard() {
   const [postItems, setPostItems] = useState([]);
@@ -96,16 +97,21 @@ function FreeBoard() {
             style={{ height: "750px", overflowY: "auto" }}
         >
           {postItems.map((postItem, index) => (
-              <CommunityPostCard
-                  key={`${postItem.postId}-${index}`}
-                  id={postItem.postId}
-                  title={postItem.title}
-                  content={postItem.content}
-                  name={postItem.memberInfo.nickname}
-                  job={"IOS Developer"}
-                  count={postItem.countInfo}
-                  img={postItem.imgUrl}
-              />
+              <>
+                <CommunityPostCard
+                    key={`${postItem.postId}-${index}`}
+                    id={postItem.postId}
+                    title={postItem.title}
+                    content={postItem.content}
+                    name={postItem.memberInfo.nickname}
+                    job={"IOS Developer"}
+                    count={postItem.countInfo}
+                    img={postItem.imgUrl}
+                />
+                <Comment
+                  postId={`${postItem.postId}`}
+                />
+            </>
           ))}
           {loading && <BarLoading />}
           {!hasMore && <EndMessage>모든 게시글을 불러왔습니다.</EndMessage>}
